Sensor verbindet und gibt Daten zurück

+++ BLETutorialYt-FinalApp/app/src/main/java/com/example/blePowerMeter/data/SensorResult.kt
@@ -1,7 +1,7 @@
 package com.example.blePowerMeter.data
 
 data class SensorResult(
-    val force:Float,
+    val force:Int,
     val angle:Float,
    val acceleration:Float,
    val connectionState: ConnectionState
